About to revert to the [d47bc088c5] state. After further thought, I have decided that the oly object shall hold the master resource.  The ICU documentation points out that "all pointers returned from the various ures_getXXX calls on this particular bundle should be considered invalid henceforth. all pointers returned by the closing of a superior object should result in all inferior pointers being null and void."
Thus, since all pointers to subresources would be void, it would not make sense to attach the errors only to state, which really is all state needs.  It would make more sense to attach the primary resource file to oly (which is intended to be unique) then attach subresources to state as needed, so state can split up into several threads.